[llcpp] Support encoding to multiple wire formats
Currently LLCPP does not provide a way to specify which wire format to
encode into. The wire format is determined simply by whether the message
is transactional or not. This CL adds support for specifying the wire
format version used in encoding.

The motivation for doing this is to make it possible to migrate
persistent encoding over to v2 piecemeal.

Work on FIDL at rest will introduce new APIs for persistent FIDL.
As a result, these APIs will soon be marked unstable and should
not be used out of tree.

Bug: 82681
Change-Id: I8d8127dcb9f4b199f31438129f9539fecae02ea3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639999
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>