[netstack3][reassembly] Make tests more flexible
Make the IP packet fragment reassembly tests more flexible so that we
can test receiving packets which have nonstandard values. This allows us
to write a test which mimics a sequence of received packets which were
found through fuzzing to cause a panic. This commit adds this test, but
marks it as `should_panic`. A future commit will be needed to debug the
issue, fix it, and remove the `should_panic` annotation.

Bug: 92831
Change-Id: I5fda32e16fd4b485259d0fd294e1d9123d2996f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641021
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>