[zxcrypt] Issue BLOCK_OP_FLUSH on Shred
Shreds really do need to make it out to the underlying storage, and we
can't rely on anything else ensuring that our writes are made durable,
so we should request an explicit flush before returning.  It also makes
sense to ask for one as we format a volume, to ensure that both birth
and death of a volume are durable.

Bug: 92790
Change-Id: Id51da9c575dbd1333bda9564b433e03a9adff399
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640768
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>