[netstack3] Use a single concrete IpSock type
Previously, we allowed testing code to use a different type for IP
sockets. This genericity turned out to be unnecessary, and caused
problems when trying to contextify the IP layer. In this commit, we
hard-code the `IpSock` type everywhere, and get rid of the testing-only
`DummyIpSock` type.

Change-Id: I451835d299d9bcc8d97d834a9afb0835e9880e4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585921
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Fuchsia-Auto-Submit: Joshua Liebow-Feeser <joshlf@google.com>