[network/tests] Fix flake
...in add_remove_address_on_loopback.

See comments inline.

On Netstack2, the DAD resolution event (which the interface watcher
depends on to update its view of interface properties) is handled
asynchronously w.r.t. the add address operation so the new IPv6 address
may not be observed in the initial "Existing" event; it may be observed
in a "Changed" event.

This is not an issue on Netstack3 as we will observe the address in an
"Existing" event right away as we do not perform DAD and Netstack3 is
queried for all interface states on watcher creation (no events happen
asynchronously).

Multiply: netstack-fidl-integration-test: 100

Change-Id: Ib1ac5493547106736861ed41e9b58ed897c4416d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641941
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>