[outgoing_directory] Introduce prototype library
This CL introduces a new version of the `service::OutgoingDirectory`
with certain differences. First, this class will live in the
`component_llcpp` namespace. Second, it will use `//sdk/lib/svc`
as a backend instead of the VFS static library. By swapping out
the backend from a static C++ library to a C shared library,
the space costs incurred by using this library will be minimized.

This library will be iterated upon until it's at parity with the
`service::OutgoingDirectory` library. Once we've reached that point,
clients of that library will be migrated to this one, and we'll
deprecate and remove `service::OutgoingDirectory`.

Bug: 89036
Bug: 92898

Change-Id: I8f0ba17f32aaa61b3522ae7f98c65a99c960abb8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/636566
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>