[virtualization] Provides tools to wait for guest pty
It is not sufficient to wait for the guest console to Start before
interracting with a guest over serial as, for example with a Linux
guest, the pty might not be started and so input might go to a
black hole. The EnclosedGuest framework solved this internally by
having repeatedly echoing until the pty was definitely up.

This moves a version of the EnclosedGuest helper into the
GuestConsole so it can be used by other test frameworks, and clarifies
that comments on the Start method to indicate its limitations.

Bug:86513

Change-Id: I76f01651a0e5099633710d36392346c36211a994
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642041
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>