[kernel][vm] Retain backlink when pinning
When a pager-backed page is moved to the Wired queue on being pinned,
retain the object backlink and offset. This allows us to validate the
object backlink irrespective of the queue the page is in, as long as it
is owned by the VmCowPages object. The vm_page_t's object field can be
shared by a stack owner if it is loaned, so this enables us to check the
backlink when that is not true and the page is owned by a VmCowPages.

Test: KernelUnittests

Bug: 92905
Change-Id: Ib7f1fca00d8d95265161c8c99e87a6f20977122e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641396
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>