[bt][audio-device-output] FrameVmo index polling
Reorient the FrameVmo polling interface to be based on frames, and not
on timestamps.  This recasts the min_duration as minimum frame count
returned as well.  Frame indexes start at zero when the buffer starts.

This means a client of FrameVmo only need track the index of the last
frame index that it received, and request all frames after that frame.

This allows accurate calculation of the time when the minimum amount of
frames will be available. Timing can be disabled by passing a zero frame
minimum, which will return an empty set of frames if none are available.

Eliminates the need for registering a separate start waker, instead
waking the waker used to poll for frames when the buffer is started.

Update the sources.rs which previously asked for more audio in frames.
This update makes the audio pipeline shorter and therefore a smaller
min_duration is required.

Bench testing suggests that this update reduces Bluetooth A2DP source
delay by around 100 milliseconds.

Bug: 73682
Test: updated tests
Test: fx test fuchsia-audio-device-output-tests
Test: fx test audio_driver_tests
Test: stream a few hours of audio with no noticable skips
Test: external speaker and test delay using audio sync YouTube video
      with P6 Pro camera
Change-Id: I49b652eb8867a9129fec2034c78ed87f3497a494
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/610882
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>