[rust] Remove unsupported target features
Apparently rustc doesn't recognize armv8-a or simd /shrug

We get a lot of this stdout spam:
```
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
[18801/31788] RUST obj/garnet/lib/rust/io_util/libio_util.rlib
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
```

Originally I thought this is necessary to get Rust auto vectorization to work.
In hindsight I'm not sure it's necessary to specify +simd.
The godbolt links here just say aarch64 which I guess is already implied?
https://www.nickwilcox.com/blog/autovec

Change-Id: Ifc2672730b2d902d003ea8dd1af1758744df2c90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642001
Fuchsia-Auto-Submit: Shai Barack <shayba@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>