[netstack3][fuzz] Generate correct UDP packets
The UDP packets being generated were using arbitrary IPv4 addresses to
produce the UDP checksum, instead of using the address from the IP
header of the outer packet, which means that most generated UDP packets
were invalid. To solve this, this commit changes the creation of UDP
packet builders to be done in tandem with the IP packet builders.

Change-Id: I2a26aad25e51925f5b24b1ac56de63d9eb5f8c64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642202
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>