[microbenchmarks] Build a host (Linux) version of fuchsia_microbenchmarks
Build a Linux version of fuchsia_microbenchmarks containing a subset
of the test cases that build on Linux without modification.

This also gets built for Mac.

Once https://fxbug.dev/90767 is implemented, we will be able to enable
running this as a performance test on NUCs on Infra, collecting
performance results.  For now, we only run it in unit test mode, which
gets run on the builders "core.x64-release" and
"core.x64-host_test_only-mac".

Bug: 93071
Change-Id: I9472b46d98cb6e8d29d34bac19f34d70bbede4f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641610
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>