[lowpanctl] Added support for `dataset` command.
The dataset command can be used to print out the current active
dataset or to change the current active dataset.

This CL also includes additional changes to the openthread_rust crate
to allow for pretty-printing of the operation dataset with the
`--format rust` option.

Change-Id: I723e3b7464624c189572cc2fbe96edb2b6e9bdbc

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642310
Reviewed-by: Jiaming Wang <jiamingw@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>