[component_hub] Timeout on `runtime` dir exploration
The `exec/runtime` directory is assumed to open or return
PEER_CLOSED. There is a third option: the directory hasn't
been served by the runner. In that case, the library will
hang waiting for a response.

To fix this, add a timeout to `runtime` dir exploration

Fixed: 93129
Change-Id: I03ff2e5a27285335a00bf668abe2dce05e36fc35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642207
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>