[wlan] Use divergent types for WPA credentials.
This change fixes an issue in the design of data structures meant to
represent WPA authentication and credentials. The set of WPA credentials
defined in FIDL (PSKs and passphrases) were assumed to be shared across
all versions of WPA, but that is not true. In particular, WPA3
explicitly does not support PSKs. To fix this, divergent types have been
introduced that are specific to versions of WPA. To make reading
authenticators more ergonomic, credentials are exposed by methods using
general types.

As this change introduces complexity, the `wpa` module has been
refactored into several sub-modules.

Change-Id: I6e153f958048bd7564f0ff813a4a619f92423411
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638627
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: Sean Olson <seanolson@google.com>