[fidl][llcpp] Change SDK back to experimental
The category of the SDK was unintentionally set to partner in this
CL - fxrev.dev/609051. This sets it back to its original state.

Note - with category 'experimental' the ddk sdk would not compile
because of level mismatch. Until this is resolved, make a local
change to set category to partner.

Change-Id: I5293041a5b6f80adb009435dad5513a1c6bbbaeb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643042
Reviewed-by: Pascal Perez <pascallouis@google.com>
Fuchsia-Auto-Submit: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>