[hypervisor][x86] Dump EPT misconfiguration info
Dump the MMU flags of the PTE when there is an EPT misconfiguration.

These are the translated MMU flags. If this doesn't provide enough
information, we will have to modify the ArchAspace code to gain access
to the raw PTEs.

Bug: 90615
Change-Id: I09e5f624eab54c719d70a72ab035c4a4976f5b59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641964
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>