[graphics][compute][tools][target_archive] Use a `source_set()` instead of toolchain_utils_action("cc")
Bug:92735 describes why the arm64 and x64 Mac builds were failing
after http://fxrev.dev/639233.

The `target_archive.gni` script has an action that compiles a
pure-data C structure and then dumps the binary with objdump.

This has worked well for many months but was revealed to be broken on
Macs by http://fxrev.dev/639233 which forced Starnix to not statically
link the Vulkan loader.

One suggestion was to provide the proper --target and --sysroot
switches to "cc" (clang) but that trips hermeticity checks when the
target's C include files are accessed.

The alternative is to:

 1. Use a standard `source_set()` to compile the structure

 2. Explicitly construct the name(s) of the `.o` outputs.  The benefit
    being that all of the platform-specific switches are handled by
    the `source_set()`.

Testability: Run optional core.{arm64,x64}-debug-mac-build_default
tryjobs as well as CQ dry run.

Change-Id: If1c7c6b698de96ecbfe6a271edbc3c83fe9cfcc2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640571
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: Allan MacKinnon <allanmac@google.com>