[kernel][x64] Always attempt to report early boot benchmarks.
Always attempt to report early boot benchmarks on x64, even when we
end up selecting HPET or PIT as our reference timer.

During physboot, and during the early phases of kernel initialization,
we take a number of EarlyTicks timestamps (always sampled from TSC) as
references for how long it took to get through various stages of the
kernel boot sequence. Later on, these values are reported (via
kcounters) for various benchmarks that we collect and track, but need
to be reported as values on the ticks timeline, which may end up using
a different reference clock than TSC on x64 (HPET and PIT being the
other valid choices).

Prior to this CL, we simply reported 0 of all of these values,
basically to indicate "I don't know".  After this CL, we make a bit
more of an effort.  During clock selection, if we choose either HPET
or PIT as our reference, we take a single correspondence pair between
the two, and set up an affine transformation which maps from the
EarlyTicks raw TSC based timeline, to the selected ticks timeline
(converting units in the process).

Assuming that the TSC didn't have its frequency changed at any point
during the boot process, and that the TSC and the reference clock
exist in the same clock domain (and are not drifting relative to each
other), the transformation we set up should end up being a fairly
accurate mapping from EarlyTicks to our chosen ticks timeline, even
though the reference clock we selected is not the same clock as the
TSC used for the collected EarlyTicks timestamps.

Tests: Manual spot checks of the reported benchmarks after forcing
selection of TSC, HPET, and PIT using boot arguments.

Change-Id: I0f6265af26a1b26c2e2b37356c4f4cc22a670e89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637019
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: John Grossman <johngro@google.com>