[driver_runtime] Refactor for implementing fdf_dispatcher_destroy()
The fdf_dispatcher_destroy() implementation will be in the next CL.
fdf_dispatcher_destroy() will need to cancel all waiting ChannelReads.

From the user's perspective, the dispatcher is registered to a
ChannelRead as soon as they call Begin(), however at that point
we have only queued it to the channel, not the dispatcher.
This change makes dispatchers aware of ChannelReads at the same
time.

This means the Channel needs to notify the dispatcher as soon as the
client calls ChannelRead::Begin(), however we don't want it to be
queued to run on the async loop yet.

Previously the Channel would call Dispatcher::QueueCallback()
when it had a pending ChannelRead wait and pending write messages.
I have split it into 2 functions:
1) Dispatcher::RegisterCallbackWithoutQueueing() and
2) Dispatcher::QueueRegisteredCallback().

Multiply: fuchsia-pkg://fuchsia.com/driver_runtime_tests#meta/driver_runtime_tests.cm : 100

Bug: 87840
Change-Id: I6bba1d21cfd19e9c1801c754f0db66690695738a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637671
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>