[routing] Allowlist detection does not use InstancedExtendedMoniker
The source_moniker stored in `CapabilityAllowlistKey` now refers to an
ExtendedMoniker that does not contain instance ids in the moniker for
the ComponentInstance variant.

`verify_routes_exceptions_allowlist.json5` is modified so that the
`component_instance.path` does not contain an `instance` field to
correspond with the non-instanced moniker.

The reason for this change is to reduce the unnecessary uses of
instanced monikers in the public API space.

Bug: 87407
Change-Id: I8c48d69c95528dce889f1ed5596ffbe38dacf351
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639231
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>