[tools][fuzz] Rewrite testcase paths in libFuzzer output
As a followup to fxrev.dev/642353, which rewrote incoming path arguments
from data/ to tmp/, this CL rewrites testcase paths back to data/ in the
libFuzzer output when running a testcase repro. This is not referenced
by ClusterFuzz during normal operation, but it is expected by one of the
integration tests.

This also fixes a bug in handling of libFuzzer args with dashes in their
right-hand side.

Change-Id: I4ab6f97937d95765c4896b99d429e9a568832058
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643286
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>