[ZBI Parser] Allow differentiating driver metadata
A ZBI item is of type driver metadata when the LSB is 0x6d. Users of the
item service will still reference this metadata by the full type, so we
must store the full type (as well as the Rust type) on each ZbiItem.

Bug: 44784

Tested:
Unit tests
CQ logs

Change-Id: I374dc8ebb663b490bbb881f9423191cf81b7aa99
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643905
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Hastings <dahastin@google.com>