[staticlints] Add new tool to wrap linters
The plan is to eventually move most of the logic from the
`tricium_analyze` recipe module into here, in the interest of moving the
linter output parsing logic in-tree so it's easier to test and modify,
and easier for non-infra engineers to integrate new linters.

For now we only use it for parsing output from Clippy (a Rust linter),
as a proof of concept.

Note that Clippy is a bit of an outlier in that it actually runs as part
of the build because it needs to actually compile the target library. So
staticlints only needs to read the Clippy output files from the build
output directory (assuming that they were generated). However, I tried
to design the `Checker` interface in a way that it will easily
generalize to linters that are run by staticlints itself.

We'll eventually want to use a similar framework for running static
analysis tools that should block CQ (e.g autoformatters). For clarity,
those tools will be run by a separate tool called `staticchecks` that
lives next to `staticlints` and uses the same framework, making it easy
to change a given linter/formatter from non-blocking to blocking or vice
versa.

Bug: 82386
Bug: 86101
Change-Id: I16061669921dc4d77cec740cffd26fdfdfdcc814
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621611
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>