[carnelian] rename and improve app context
AppContext was too generate a name for a struct whose sole
responsibility was sending messages to be handled in the executor main
loop. This change renamed it to AppSender and enhances it with a way to
use the AppSender from another thread.

Testing: button uses a thread to drive the resize messages.

Change-Id: I508c5dbe3bd2f5bbc58d7c97249ca0b9401ff00e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642441
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: Rob Tsuk <robtsuk@google.com>