[workstation] Bundle Fxfs
This doesn't enable Fxfs, it just includes it into the build so that
it's available for enablement at runtime.  Currently this is only
possible by developers, but it makes testing easier (we can enable Fxfs
on devices which are using regular builds using a flag introduced in
fxrev.dev/640357).

Bug: 92302
Bug: 92463
Change-Id: I974282e7d7c68860a1326dae1409ec57dede030f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641395
Reviewed-by: Aaron Drew <ripper@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>