[security][pkg] Centralize update URL configuration
This change centralizes the domain used for downloading updates
in security package delivery tests. Previously, the choice of name
was implicit, and based on the set of domains designated across
`pkg-resolver` configuration files. This change designates a
particular domain explicitly in a config file, and routes the
configuration information into all host and target tools/tests
that depend on the domain name.

Bug: 88453
Change-Id: I0d3892fca4383bc454b843ad70842166a11189f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643705
Reviewed-by: Allison Pearce <ampearce@google.com>
Commit-Queue: Mark Dittmer <markdittmer@google.com>