[docs] Fixing shell commands on the fx page
This CL updates the //development/build/fx.md doc shell blocks. Some of
the blocks had "$" characters before the commands the reader is meant to
copy, so when they pressed the block's copy button, the "$" was copied
as well. This meant when pasting the text, the user had to edit the
copied text.

In this change, I updated a few short command blocks to use
"posix-terminal" instead of "shell. This places a "$" before the text,
but it doesn't get copied to the user's clipboard, which makes
copy/paste more convenient. I also updated a few of the longer blocks,
which contained both code and comments, by removing the ability to copy
and paste.

Change-Id: Ib40b32a753e312a94d93f5170c6cb3cbf72dbeef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643641
Reviewed-by: Kyo Lee <kyol@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>