[kernel][lib/arch] Conditionalize hwreg::EnablePrinter use
hwreg::EnablePrinter is a nice convenience that comes with a large
runtime overhead in object size.  It's mostly used only for
build-time generator programs, and elsewhere only in test code.
So use a build-time conditional in lib/arch headers to enable
the printing support only in the places that need it.

Change-Id: I5a89754c5508d47594f8c80060d4a4d772984f3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644041
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>