[build][rust] assert rust edition is 2015 or 2018
We don't yet support rust edition 2021 in tree yet, but we didn't actually
assert that the edition value was 2015 or 2018. This allowed targets to
specify 2021, which would just fall back to 2018. This adds an assertion,
and switches all the targets that set `edition = 2021` to `edition = 2018`.

Bug: 90449
Change-Id: I5177a5ec472b0ff85a02c875f665b825eeff4a0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644263
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Joseph Ryan <josephry@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>