[forensics] Prevent exceptions from going into oblivion
OnException is a "move-only" method and any exception sent via the
method can't be reclaimed if there's an error with the server. The
exception handling protocol has been amended to include a method that
allows clients to check if the server is able to respond to requests
before they make the call to OnException and cede ownership of the
exception.

Note: A response to IsActive doesn't mean an error didn't occur between
calls to IsActive and HandleException, but it dramatically shrinks the
window during which errors can occur and not be recoverable.

Fixed: 57111
Bug: 93282

TESTED=`/boot/test/crashsvc-test`
TESED=added crash to exceptions.cm and exception was released.

Change-Id: Ia32bb478e3e12f2b2cbba4fb399a26119d34b394
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643786
Reviewed-by: Francois Rousseau <frousseau@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
API-Review: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Alex Pankhurst <pankhurst@google.com>