[terminal] Improved scroll bar
Switch to an overlay scroll bar that is hidden when
not needed. This means that the scroll bar is no
longer reducing the size of the terminal and it
provides a more immersive experience for apps
that use app-cursor mode.

The scroll bar thumb is also using alpha so text below
it is still visible. Thumb alpha is adjusted depending
on the location of the mouse and if the thumb is
grabbed for scrolling.

This also makes the terminal handle mouse and touch
input better in order to enable selections using the
mouse in the future.

Test: terminal_tests
Change-Id: I431622d5fd77ae39c72f922ce074d2327287686f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644531
Reviewed-by: Rob Tsuk <robtsuk@google.com>
Commit-Queue: David Reveman <reveman@google.com>