[kernel][vm] Test mapping protect with dirty tracking
A zx_vmar_protect that adds the WRITE permission should not overwrite
any hardware mappings that are set up as READ-only in order to mark
pages dirty on a write. zx_vmar_protect does not allow adding WRITE for
any mapping today, so this just works. If this behavior changes in the
future however, the test added in this CL will help us catch it.

Test: core-pager-writeback-test

Bug: 63989
Change-Id: Id929ce126820359639b56111a86456436e60e029
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644111
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>