[bt][usb] Fix bug processing multiple CMD packets
When multiple command packets were sent to the transport at once, the
length of the buffer accepting the packets was overwritten and not
reset, resulting in rejection of a packet and channel shutdown.

Reset the buffer size before attempting to read from the channel, and
drop messages that are too large instead.

Test: added tests; fx test bt-transport-usb-tests
Change-Id: I7d2f43b8093de313c8a18b504ce9008ca2afca55
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644421
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>