[fs_management] Add support for matching children of a device.
The paver uses fs_management to find all FVM partitions of a particular
type so that it can wipe them before writing a new one to the FVM.
Currently, it doesn't support systems with more than one FVM - since
PartitionMatches always returns the first matching partition it finds,
it may return a partition belonging to the wrong FVM.

This constraint was fine previously - zedboot does not have a ramdisk
FVM built-in - but for writing a completely new FVM from recovery, we
need to handle the case where multiple FVMs are present on a system and
we need to distinguish between them.

This CL adds support for a "parent_device" member in `struct
PartitionMatcher`, which causes fs_management to check that the
found partition is a child of the given device.

Change-Id: Ie083dbfc89aff53177a7801ef79fc79cad427124