[examples][fidl] Avoid using potentially undefined operations
All of the LLCPP server examples contain code like the following:

  std::unique_ptr server = std::make_unique<EchoImpl>();
  server->Bind(std::move(server), dispatcher, std::move(request));

The intent is to give EchoImpl ownership of itself so it can destroy
itself after the channel closes. This makes sense. The problem is the
second line, particularly:

  server->Bind(std::move(server), ...

I believe this is undefined behavior before C++17: we don't know whether
the deref of server will happen before or after the std::move. This
appears to be defined in C++17 in the expected way, as deref before
move, at least based on my reading of the following link (see rule 14).
I haven't checked the precise language in the C++14 and C++17 specs.
https://en.cppreference.com/w/cpp/language/eval_order

In any case, although the example as-written has an elegant simplicity
in some ways, we probably shouldn't rely on a potential obscure corner
of C++ in "intro to FIDL" example code.

This CL changes EchoImpl to delete itself once it is unbound. This
change is made to two example programs (server and services/server).
Markdown documentation for the first program has been updated. The
second program has no markdown documentation -- the markdown page is
just a link to the source code.

Change-Id: I42b024e28f362d5248135230d717178f5d531e4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640771
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>