[docs] Fix RFC metadata
This CL fixes entries in _rfcs.yaml to use [''] rather than []. It also
fixes the rfcid in RFC-0088, and an incorrect heading level in RFC-0063.

The reason to prefer [''] is that the Jinja2 logic in _rfc_metadata.md
checks specifically for that. If [] is used, it emits an empty table row
instead of omitting it. While we could change that logic to check for []
instead, this default would likely lead people to edit [] to [12345]
(for example), whereas it should be ['12345'].

Change-Id: I9bcc081160abd1f1e236e8a0a6dccaea3ecb7712
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644412
Reviewed-by: Pascal Perez <pascallouis@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>